Refactors handleVotes into utils.js
